Made points for user count as beers and track them :beers:

+++ annotator_supreme/controllers/database_controller.py
@@ -146,7 +146,7 @@ class DatabaseController:
                     username text,
                     password_hash text,
                     email text,
-                    points int,
+                    points double,
                     registration_date timestamp,
                     PRIMARY KEY (username)
                 )
